refactor: address PR review — move SpacePermissionError to exceptions.py
- Move SpacePermissionError from space_creator.py to exceptions.py,
  inheriting from APIError per codebase convention
- Move inline import in test_space.py to top-level import block
- Tighten dry-run assertion from `is not None` to exact value check
